sql: move MergeResultTypes from physicalplan package and unexport
Release note: None
