+++ pkg/sql/BUILD.bazel
@@ -68,6 +68,7 @@ go_library(
         "distsql_plan_ctas.go",
         "distsql_plan_join.go",
         "distsql_plan_scrub_physical.go",
+        "distsql_plan_set_op.go",
         "distsql_plan_stats.go",
         "distsql_plan_window.go",
         "distsql_running.go",
@@ -425,6 +426,7 @@ go_test(
         "descriptor_mutation_test.go",
         "distsql_physical_planner_test.go",
         "distsql_plan_backfill_test.go",
+        "distsql_plan_set_op_test.go",
         "distsql_running_test.go",
         "drop_helpers_test.go",
         "drop_test.go",
