+++ pkg/sql/physicalplan/physical_plan_test.go
@@ -8,10 +8,6 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-// This file defines structures and basic functionality that is useful when
-// building distsql plans. It does not contain the actual physical planning
-// code.
-
 package physicalplan
 
 import (
@@ -376,47 +372,5 @@ func TestProjectionAndRendering(t *testing.T) {
 		if o := strings.Join(ord, ","); o != tc.expOrdering {
 			t.Errorf("%d: incorrect ordering: '%s' expected '%s'", testIdx, o, tc.expOrdering)
 		}
-	}
-}
-
-func TestMergeResultTypes(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-	defer log.Scope(t).Close(t)
-
-	empty := []*types.T{}
-	null := []*types.T{types.Unknown}
-	typeInt := []*types.T{types.Int}
-
-	testData := []struct {
-		name     string
-		left     []*types.T
-		right    []*types.T
-		expected *[]*types.T
-		err      bool
-	}{
-		{"both empty", empty, empty, &empty, false},
-		{"left empty", empty, typeInt, nil, true},
-		{"right empty", typeInt, empty, nil, true},
-		{"both null", null, null, &null, false},
-		{"left null", null, typeInt, &typeInt, false},
-		{"right null", typeInt, null, &typeInt, false},
-		{"both int", typeInt, typeInt, &typeInt, false},
-	}
-	for _, td := range testData {
-		t.Run(td.name, func(t *testing.T) {
-			result, err := MergeResultTypes(td.left, td.right)
-			if td.err {
-				if err == nil {
-					t.Fatalf("expected error, got %+v", result)
-				}
-				return
-			}
-			if err != nil {
-				t.Fatalf("unexpected error: %s", err)
-			}
-			if !reflect.DeepEqual(*td.expected, result) {
-				t.Fatalf("expected %+v, got %+v", *td.expected, result)
-			}
-		})
 	}
 }
